% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LocalInt}
\alias{LocalInt}
\title{Local Interval Mapping}
\usage{
LocalInt(peakChr, peakPos, range = 100, phenotype.dat, pheno.name, design)
}
\arguments{
\item{peakChr}{character vector of length one. Must be one of the
  major chromosome arms in the \emph{Drosophila} genome
  ('X','2L','2R','3L',or '3R').}

\item{peakPos}{numeric vector of length one. A position in base
  pairs in the DSPR position list (every 10kb).}

\item{range}{numeric vector of length one specifying the number of
  positions to test on either side of the peak (positions are
  every 10kb). Default is 100.}

\item{phenotype.dat}{\code{data.frame} containing a column of ril
  ids (must be named patRIL) and phenotypes.}

\item{pheno.name}{a character string specifying the column name of
  the phenotype data}

\item{design}{a character string. One of either 'inbredA' or
  'inbredB' corresponding to the pA and pB set of inbred RILs.
  Other crossing designs may be supported in the future.}
}
\value{
A \code{data.frame} consisting of the chromosome, physical
  position (bp), genetic position (cM) and LOD score for each
  position.
}
\description{
\code{LocalInt} performs standard interval mapping around a given
peak for a range of positions specified by the user.
}
\author{
Elizabeth King (\email{kingeg@missouri.edu})
}

