\docType{data}
\name{positionlist_wgenetic}
\alias{positionlist_wgenetic}
\alias{poslist}
\title{Positionlist}
\format{A data frame with 4 variables.
\describe{
\item{\code{chr}}{character vector of chromosome}
\item{\code{Ppos}}{numeric vector}
\item{\code{Gpos}}{numeric vector}
\item{\code{Gaxis}}{numeric vector}
}}
\description{
  Long description
}
\keyword{datasets}

