% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSPRqtl-package.R
\docType{data}
\name{ADHdata}
\alias{ADHdata}
\title{ADH data}
\format{A data frame with 3 variables.
\describe{
\item{\code{patRIL}}{numeric vector}
\item{\code{matRIL}}{numeric vector}
\item{\code{adh}}{numeric vector}
}}
\description{
The ADH activity phenotype data.
}
\keyword{datasets}
