% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DSPRpeaks}
\alias{DSPRpeaks}
\alias{print.peaks}
\title{Find and Summarize QTL}
\usage{
DSPRpeaks(qtldat, method, threshold, LODdrop, BCIprob)
}
\arguments{
\item{qtldat}{An object of class gscan. Output from
  \code{\link{DSPRscan}}.}

\item{method}{a character string specifying the method to use for
  the confidence interval. Options are: 'LODdrop' calculates a LOD
  drop interval for the drop amount specified, 'BCI' calculates
  the Bayesian credible interval for the fraction specified, and
  'both' calculates and returns both.}

\item{threshold}{numeric vector of length one consisting of the
  signficance threshold. Default is 6.8 for inbred designs and
  10.1 for the ABcross. Use \code{\link{DSPRperm}} to get a
  threshold specific to a given dataset.}

\item{LODdrop}{numeric vector of length one consisting of the LOD
  drop to be used when using method 'LODdrop'. Default value is 2
  which approximates a 95\% confidence interval for the inbred
  designs. Users of the ABcross design should consider using a
  larger LOD drop.}

\item{BCIprob}{numeric vector of length one consisting of the
  nominal Bayes fraction. Default value is 0.95.}
}
\value{
A list of class \code{peaks} containing a list for each
  significant peak, each containing:
\item{threshold}{the specified signficance threshold}
\item{peak}{A single row \code{data.frame} with the chromosome,
physical position (bp), genetic position (cM) and LOD score for
the peak}
\item{LODdrop}{the specified LOD drop for the support interval}
\item{CI}{the upper and lower bounds of the confidence interval}
\item{founderNs}{the number of RILs with each founder genotype at
the peak}
\item{geno.means}{the estimated means and standard errors for each
founder genotype}
\item{perct.var}{the percent variation explained by the QTL}
\item{entropy}{the proporiton of missing information (entropy)}
}
\description{
\code{DSPRpeaks} takes output from \code{\link{DSPRscan}}. Locates
and summarizes QTL peaks.
}
\author{
Elizabeth King (\email{kingeg@missouri.edu})
}

