\docType{package}
\name{DSPRqtl-package}
\alias{DSPRqtl}
\alias{DSPRqtl-package}
\title{DSPRqtl}
\description{
  Analysis of DSPR phenotypes
}
\details{
  \tabular{ll}{ Package: \tab DSPRqtl\cr Type: \tab
  Package\cr Version: \tab 1.0-4\cr Date: \tab
  2012-04-26\cr License: \tab GPL-2\cr LazyLoad: \tab
  yes\cr LazyData: \tab yes\cr }
}
\author{
  Elizabeth King
}
\keyword{package}

