% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSPRqtl-package.R
\docType{package}
\name{DSPRqtl-package}
\alias{DSPRqtl-package}
\alias{DSPRqtl}
\title{DSPRqtl}
\description{
Analysis of DSPR phenotypes
}
\details{
\tabular{ll}{
  Package: \tab DSPRqtl\cr
  Type: \tab Package\cr
  Version: \tab 2.0-4\cr
  Date: \tab 2015-06-21\cr
  License: \tab GPL-2\cr
  LazyLoad: \tab yes\cr
  LazyData: \tab yes\cr
}
}
\author{
Elizabeth King
}
\keyword{package}
