% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findQTL}
\alias{findQTL}
\title{Find QTL}
\usage{
findQTL(qtldat, threshold, design)
}
\arguments{
\item{qtldat}{\code{data.frame} of chromosome, position, and LOD
  scores (column names chr,Ppos,Gpos,LOD). List element LODscores
  from \code{\link{DSPRscan}}.}

\item{threshold}{numeric vector of length one consisting of the
  signficance threshold. Default is 6.8 for inbred designs and
  10.1 for the ABcross. Use \code{\link{DSPRperm}} to get a
  threshold specific to a given dataset.}

\item{design}{a character string. One of either 'inbredA',
  'inbredB', or 'ABcross' corresponding to the pA and pB set of
  inbred RILs or the pA-pB cross design.}
}
\value{
A \code{data.frame} consisting of the chromosome, physical
  position (bp), genetic position (cM) and LOD score for each
  significant peak.
}
\description{
\code{findQTL} finds peak LOD scores that exceed the given
threshold (identifies signficant QTL).
}
\author{
Elizabeth King (\email{egking@uci.edu})
}

