% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{positionlist_wgenetic}
\alias{positionlist_wgenetic}
\alias{poslist}
\title{Positionlist}
\format{A data frame with 4 variables.
\describe{
\item{\code{chr}}{character vector of chromosome}
\item{\code{Ppos}}{numeric vector}
\item{\code{Gpos}}{numeric vector}
\item{\code{Gaxis}}{numeric vector}
}}
\description{
List of regularly spaced positions every 10kb used for data
analysis of DSPR data. chr is the chromosome arm, Ppos is the
position in base pairs, Gpos is the position in centiMorgans, and
Gaxis is used for plotting the entire genome on a single axis.
}
\keyword{datasets}

